backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
psycopg2-binary==2.9.1
requests==2.26.0
python-dotenv==0.19.1
gunicorn==20.1.0